Corrected parameter names - diameter/radius
In a previous commit, parameter names were updated from *size* to *diameter* . In this commit, those parameters were either verified to be diameters or updated to be radii. These changes in populate_model.R result in changes to the Number_Of_Particles_Factor calculation as well as the assignment of particle radii in MoBi. Note that the deposition fractions still match the Boger & Wigstrom (2018) MATLAB code. The inhalation_model_user_guide was updated to v1.3 with changes to the parameter names, updated images for the tutorial for step 7 and the results section, and correction of typos.