Added solubility in ELF
v1.2 of the inhalation model incorporates an additional parameter "Solubility in epithelial lining fluid" in the molecule building block. The default formula for this parameter is to copy the value for "Solubility at reference pH" for the molecule. This new parameter is used in the formula "PARAM_ParticlesReleaseRate" which is used in the parameter "Release rate of the tablet" which can be found in the "ParticlesDrugRelease_*" transports within the "Events - Standard" building block.